build-from-source: skip test checking for gtk-doc issue
For libnm, we use opaque types. gtk-doc has/had an issue parsing
this code, and generates suboptimal documentation.

There is a merge request against gtk-doc to address that [1].

However, there is also a `make check` test, which tries to determine
whether gtk-doc is suitable [2].

When building for beaker, we don't need this check. Also, because likely
beaker is not up to the task.

[1] https://gitlab.gnome.org/GNOME/gtk-doc/merge_requests/2
[2] https://github.com/NetworkManager/NetworkManager/pull/196