build-from-source: fix build failure due to netconfig path
During the test build we enabled "--with-netconfig=yes".

Since commit "5b36585a3d build/autotools: fail configure if
netconfig/resolveconf tool is not found", when specifying
"--with-netconfig=yes" the user is required to have netconfig
installed (so that the path can be detected). Otherwise it fails
with

    checking for netconfig... no
    configure: error: cannot find netconfig in path. Set the path explicitly via --with-netconfig=PATH.

The correct way is to explicitly specify the path. In that
case, it's OK that the file doesn't actually exist.