libnm: don't skip NMObject:path from documentation and introspection
The D-Bus path is a useful property, also exposed via
nm_object_get_path() function. Don't mark it to be
skipped from documentation and introspection.

This was changed in "1f5b48a59e libnm: use the o.fd.DBus.ObjectManager
API for object management" and was an API breakage of libnm.
This also caused a bug in gnome-shell [1].

[1] https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/240/diffs
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1628263

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/merge_requests/16

Fixes: 1f5b48a59eb46c40cb10bf4381b2b21a19a9f471